Fix sphinx doc problem.

+++ scripts/travis_after_success.sh
@@ -12,14 +12,14 @@ if [[ ${TRAVIS_BRANCH} == "master" ]] && [[ ${TRAVIS_PULL_REQUEST} == "false" ]]
   # Decrypt the private key.
   openssl aes-256-cbc -k "$travis_key_password" -md sha256 -d -a -in travis_key.enc -out ./travis_key
 
-  if [ -s ./travis_key ]
+  if [[ -s ./travis_key && $? == 0 ]]
   then
 	chmod 400 ./travis_key
   	echo "Host github.com" > ~/.ssh/config
   	echo "  IdentityFile $(pwd)/travis_key" >> ~/.ssh/config
 
-  	git clone --bare ${mirror_from} forte_bare
-  	if cd forte_bare; then
+  	git clone --bare ${mirror_from} bare
+  	if cd bare; then
     	git push --mirror ${mirror_to}
   	else
    		echo "Cannot cd into forte_bare, clone may be unsuccessful."
